save old files

+++ scrapper.go
@@ -3,14 +3,12 @@ package main
 import (
 	"bufio"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
-	"regexp"
-	"strconv"
 	"strings"
-
 	"web-scrapper/utility"
 
 	"github.com/gocolly/colly"
@@ -25,6 +23,7 @@ type Game struct {
 
 func main() {
 	checkLinks()
+	checkHistory()
 }
 
 func initCollectro(c *colly.Collector, games *[]Game) {
@@ -48,7 +47,7 @@ func initCollectro(c *colly.Collector, games *[]Game) {
 			game.Name = el.ChildAttr("div.relative.hoverable-box.d-flex.flex-wrap.flex-align-center.game-item.cta-full.item.game-deals-item.game-list-item.keep-unmarked-container", "data-game-name")
 			game.ShopName = el.ChildAttr("div.relative.hoverable-box.d-flex.flex-wrap.flex-align-center.game-item.cta-full.item.game-deals-item.game-list-item.keep-unmarked-container", "data-shop-name")
 			price := el.ChildText(".price-inner.game-price-current")
-			game.Price = parsePrice(price)
+			game.Price = utility.ParsePrice(price)
 			link := el.ChildAttr(".full-link", "href")
 			game.Link = el.Request.AbsoluteURL(link)
 			*games = append(*games, game)
@@ -79,29 +78,36 @@ func checkLinks() {
 		gameJSON, err := json.Marshal(games)
 		utility.Check(err)
 
-		filename := strings.Trim(baseURL[22:], "/") + ".json"
-		f, err := os.Create(filename)
-		utility.Check(err)
-		defer f.Close()
+		filename := "./data/" + strings.Trim(baseURL[22:], "/") + ".json"
+
+		if _, err := os.Stat(filename); errors.Is(err, os.ErrNotExist) {
+			f, err := os.Create(filename)
+			utility.Check(err)
+			defer f.Close()
+		} else if err == nil {
+			e := os.Rename(filename, "./data/"+strings.Trim(baseURL[22:], "/")+"_old"+".json")
+			utility.Check(e)
+			f, err := os.Create(filename)
+			utility.Check(err)
+			defer f.Close()
+		} else {
+			utility.Check(err)
+		}
+
 		err = ioutil.WriteFile(filename, gameJSON, 0644)
 		utility.Check(err)
 	}
 }
 
-// parse price into list of floats
-func parsePrice(price string) []float64 {
-	r, _ := regexp.Compile("[+-]?([0-9]*[,])?[0-9]+")
-
-	var prices []float64
-
-	fmt.Println("Price: ", r.FindAllString(price, -1))
-	for _, p := range r.FindAllString(price, -1) {
-		p = strings.Replace(p, ",", ".", -1)
-		if s, err := strconv.ParseFloat(p, 32); err == nil {
-			prices = append(prices, s)
-		}
+func checkHistory() {
+	dat, err := os.Open("games.txt")
+	utility.Check(err)
+	defer dat.Close()
+	scanner := bufio.NewScanner(dat)
+	for scanner.Scan() {
+		//filename := strings.Trim(scanner.Text()[22:], "/") + ".json"
+		//content, err = ioutil.ReadFile("./data/" + filename)
+		utility.Check(err)
 	}
-
-	return prices
 
 }
